# business/catalogue-service-micro/src/main/resources/api.yml
...
swagger: '2.0'
...
info:
  version: "0.0.1"
  title: Catalogue Service API

# Describe your paths here
paths:
...
  /catalogue/{standard}:
    post:
      operationId: addXMLCatalogue
      summary: Add catalogue
      description: Adds the catalogue represented in XML format. The serialized catalogue should be compliant with the specified standard
      consumes: [application/json, application/xml]
      produces: [application/json, application/xml]
      parameters:
      - in: path
        name: standard
        required: true
      - in: body
        name: serializedCatalogue
        required: true
      responses:
        201:
          description: Saved catalogue is returned
        400:
          description: Invalid content type header or invalid standard
    put:
      operationId: updateJSONCatalogue
      summary: Update JSON catalogue
      description: Updates the catalogue represented in JSON serialization. The serialization should be compliant with the default standard, which is UBL
      consumes: [application/json]
      produces: [application/json]
      parameters:
      - in: body
        name: catalogueJson
        required: true
      responses:
        304:
          description: Updated catalogue is returned
        400:
          description: Invalid standard or invalid catalogue serialization
        501:
          description: An unsupported standard is provided
  /catalogue/{standard}/{uuid}:
    get:
      operationId: getCatalogue
      summary: Returns the catalogue with the given standard and uuid
      description: Retrieves the catalogue for the given standard and uuid.
      produces:
      - application/json
      parameters:
      - in: path
        name: standard
        description: Standard of the requested catalogue
        required: true
        type: string
      - in: path
        name: uuid
        description: Identifier of the catalogue
        required: true
        type: string
      responses:
        200:
          description: The catalogue specified with the uuid is returned
        204:
          description: No catalogue exists for the given uuid
        400:
          description: Invalid standard
    delete:
      operationId: deleteCatalogue
      summary: Delete catalogue
      description: Deletes the catalogue for the given standard and uuid.
      parameters:
      - in: path
        name: standard
        description: Standard of the catalogue to be deleted
        required: true
        type: string
      - in: path
        name: uuid
        description: Identifier of the catalogue to be deleted
        required: true
        type: string
      responses:
        200:
          description: Catalogue deleted, if one exists
        400:
          description: Invalid standard
...
  /catalogue/{catalogueUuid}/catalogueline:
    post:
      operationId: addCatalogueLine
      summary: Add catalogue line
      description: Adds the provided catalogue line
      consumes: [application/json]
      produces: [application/json]
      parameters:
      - in: path
        name: catalogueUuid
        description: Unique identifier of the catalogue to which the provided catalogue line will be added
        required: true
        type: string
      responses:
        201:
          description: Returned created catalogue line
        400:
          description: Failed to parse the provided catalogue line
    put:
      operationId: updateCatalogueLine
      summary: Update catalogue line
      description: Adds the provided catalogue line
      consumes: [application/json]
      produces: [application/json]
      parameters:
      - in: path
        name: catalogueUuid
        description: Unique identifier of the catalogue containing including the provided catalogue line
        required: true
        type: string
      responses:
        200:
          description: Returned updated catalogue line
        400:
          description: Failed to parse the provided catalogue line
  /catalogue/{catalogueUuid}/catalogueline/{lineId}:
    get:
      operationId: getCatalogueLine
      summary: Get catalogue line
      description: Retrieves the catalogue line specified with the lineId parameter
      produces: [application/json]
      parameters:
      - in: path
        name: catalogueUuid
        description: Unique identifier of the catalogue including the requested catalogue line
        required: true
        type: string
      - in: path
        name: lineId
        description: Identifier of the catalogue line
        required: true
        type: string
      responses:
        200:
          description: Returned requested catalogue line
        204:
          description: There does not exist a catalogue line for the given parameters
    delete:
      operationId: deleteCatalogueLine
      summary: Delete catalogue line
      description: Deletes the specified catalogue line, if exists
      parameters:
      - in: path
        name: catalogueUuid
        description: Unique identifier of the catalogue including the catalogue line to be deleted
        required: true
        type: string
      - in: path
        name: lineId
        description: Identifier of the catalogue line to be deleted
        required: true
        type: string
      responses:
        200:
          description: Deleted catalogue line
  /catalogue/category:
    get:
      operationId: getCategories
      summary: Get categories
      description: Retrieves a list of categories. This method takes either a list of names or a list (category id/taxonomy id) pairs. When the category names are provided, they are looked for in all managed taxonomies. For the other option, (i.e. get by id option) the taxonomy id is already provided along with the category id. See the examples in parameter definitions.
      produces: [application/json]
      parameters:
      - in: query
        name: categoryNames
        description: List of names. Example "wood, mdf"
        required: false
        type: string
      - in: query
        name: taxonomyIds
        description: List of taxonomy ids. IDs must indicate the taxonomies containing the corresponding category in the categoryIds parameter. Example "eClass, eClass"
        required: false
        type: string
      - in: query
        name: categoryIds
        description: List of category ids. Example "0173-1#01-BAC439#012,0173-1#01-AJZ694#013"
        required: false
        type: string
      responses:
        200:
          description: Returned categories matched with the specified parameters
        400:
          description: None of the category names or (taxonomyId/categoryId) pairs are provided; or number of elements in taxonomy id and category id lists do not match
...